requirements.txt
Flask==0.10.1

# Application server for both development and production.
gunicorn==19.4.5
uwsgi

# Testing and static analysis.
pytest==2.9.1
pytest-cov==2.2.1
flake8==2.5.4

# CLI.
Click==6.4

# Data and workers.
psycopg2
Flask-SQLAlchemy==2.1
Flask-Migrate==2.1.1
redis==2.10.5
celery==3.1.23

# Forms.
Flask-WTF==0.9.5
WTForms-Components==0.9.7

# Utils.
fake-factory==0.5.7

# Extensions.
flask-debugtoolbar==0.10.0
Flask-Mail==0.9.1
Flask-Login==0.3.2
